update and add api copy message group, message bag

See merge request smirnovbui/instagram-chatbot!131

+++ app/controllers/api/v1/message_managements/message_bags_controller.rb
@@ -9,6 +9,7 @@ class Api::V1::MessageManagements::MessageBagsController < ApplicationController
   def show
     message_bag = MessageBag.find_by(id: params[:id])
     return render json: {code: 2, message: "Cannot find message bag"} if message_bag.blank?
+    return render json: {code: 2, message: "User can't permission"} if message_bag.message_group.user_id != current_user.id
     messages = Message.where(message_bag: message_bag)
     render json: {code: 1, data: {message_bag: message_bag, messages: messages}}
   end
@@ -16,14 +17,27 @@ class Api::V1::MessageManagements::MessageBagsController < ApplicationController
   def update
     message_bag = MessageBag.find_by(id: params[:id])
     return render json: {code: 2, message: "Cannot find message bag"} if message_bag.blank?
-    message_bag.update message_bag_params
+    return render json: {code: 2, message: "User can't permission"} if message_bag.message_group.user_id != current_user.id
     render json: {code: 1, data: message_bag}
   end
 
   def destroy
     message_bag = MessageBag.find_by(id: params[:id])
     return render json: {code: 2, message: "Cannot find message bag"} if message_bag.blank?
+    return render json: {code: 2, message: "User can't permission"} if message_bag.message_group.user_id != current_user.id
+    return render json: {code: 2, message: "Cannot delete message bag"} unless MessageBagForm.new(message_bag, current_user).check_delete
     if message_bag.destroy
+      render json: {code: 1, message: "Success!"}
+    else
+      render json: {code: 2, message: "Something went wrong!"}
+    end
+  end
+
+  def copy
+    message_bag = MessageBag.find_by(id: params[:id])
+    return render json: {code: 2, message: "Cannot find message bag"} if message_bag.blank?
+    return render json: {code: 2, message: "User can't permission"} if message_bag.message_group.user_id != current_user.id
+    if CopyObject::MessageManager.new(message_bag, "bag").call
       render json: {code: 1, message: "Success!"}
     else
       render json: {code: 2, message: "Something went wrong!"}
